+++ priv/repo/migrations/20251122020602_add_shop_domain.exs
@@ -35,7 +35,8 @@ defmodule Medishop.Repo.Migrations.AddShopDomain do
             name: "orders_location_id_fkey",
             type: :uuid,
             prefix: "public"
-          ), null: false
+          ),
+          null: false
 
       add :user_id,
           references(:users,
@@ -43,7 +44,8 @@ defmodule Medishop.Repo.Migrations.AddShopDomain do
             name: "orders_user_id_fkey",
             type: :uuid,
             prefix: "public"
-          ), null: false
+          ),
+          null: false
     end
 
     create unique_index(:orders, [:order_number], name: "orders_unique_order_number_index")
@@ -68,7 +70,8 @@ defmodule Medishop.Repo.Migrations.AddShopDomain do
             name: "order_items_order_id_fkey",
             type: :uuid,
             prefix: "public"
-          ), null: false
+          ),
+          null: false
 
       add :product_id,
           references(:products,
@@ -76,7 +79,8 @@ defmodule Medishop.Repo.Migrations.AddShopDomain do
             name: "order_items_product_id_fkey",
             type: :uuid,
             prefix: "public"
-          ), null: false
+          ),
+          null: false
     end
 
     create table(:carts, primary_key: false) do
@@ -96,7 +100,8 @@ defmodule Medishop.Repo.Migrations.AddShopDomain do
             name: "carts_location_id_fkey",
             type: :uuid,
             prefix: "public"
-          ), null: false
+          ),
+          null: false
     end
 
     create unique_index(:carts, [:location_id], name: "carts_unique_cart_per_location_index")
@@ -120,7 +125,8 @@ defmodule Medishop.Repo.Migrations.AddShopDomain do
             name: "cart_items_cart_id_fkey",
             type: :uuid,
             prefix: "public"
-          ), null: false
+          ),
+          null: false
 
       add :product_id,
           references(:products,
@@ -128,7 +134,8 @@ defmodule Medishop.Repo.Migrations.AddShopDomain do
             name: "cart_items_product_id_fkey",
             type: :uuid,
             prefix: "public"
-          ), null: false
+          ),
+          null: false
     end
 
     create unique_index(:cart_items, [:cart_id, :product_id],